Expose builtin for trace.
This exposes the trace region functionality.

It is does not correctly impose constraints on callers that they cannot
keep references into this region.  Thus it
can lead to dangling pointers.

Type system work required to fix this API.

The example illustrates how we might use trace on a per-data-structure
level.