Correctly follow the calling convention on program entry.
Static methods should be called with their descriptor as their first
argument, but until now the program entrypoint was called with an
invalid receiver.

To enable this, we store Main's descriptor and selector in the program
header, instead of the offset to the entrypoint as we did previously. On
startup the VM performs a vtable lookup with this descriptor and
selector to find the right offset, like any method call would do.

Descriptor and selector count fields in the program header are switched
to 32-bit, given that all indices were 32-bit already.