Use `target_link_libraries` for pegmatite.
We used to manually add the right sources and compiler search paths, but
using `target_link_libraries` is nicer as it will do all that and also
add the right compiler definitions (eg. `-DUSE_RTTI`).

TODO: this won't actually work until we update the pegmatite submodule,
as it needs CompilerTeaching/Pegmatite#35.