Add primitive entities
Primitives are used for types which have a special meaning in the
compiler. In the future this may be used for U64, cowns, external
references, arrays, ...

Instances of primitives can only be created through special opcodes, never
with `new`. Each primitive can have a different underlying representation, and
they have no fields. They may however have both static and dynamic methods.
These methods can be implemented either in Verona code, or as builtin methods.

This change simplifies these special types, which are currently hardcoded in the
compiler, and are a step towards unifying them with regular classes.