Make the binary-tree test use `mut` nodes.
The test used to fail when run under ASan, as the mut Values were
allocated in the region of the tree root, instead of their Node's
region.

This commit changes the tree to always use `mut` nodes, allowing the
values to be allocated in that region too.

It would be neat to use `mut` nodes when storing `mut` values, and `iso`
nodes when storing `iso` values, but I can't think of an way to do this
currently, in a generic fashion.