Unify CI configuration for RT vs Lang.
Do the lang and runtime tests in the same job, on all platforms.

This has the following effects:
- Enable both Debug/Release testing of the language.
- Enable ASAN testing of the language.
- Enable runtime tests on macOS.

The Linux setup is cleaned up a bit, by using CMake flags to pick the
compiler rather than messing with update-alternatives. I was going to
add a GCC configuration, hence the CC/CXX/CXXFLAGS refactor, but runtime
tests are currently broken on GCC and it needs a non-trivial amount of
work.