Improve SFINAE code in vobject.
If for example a class C has a finaliser method with an incorrect
signature, `has_finaliser<C>` will still be true. However the
implementation of V<C> (in this case `gc_final`) would not compile.

This is better than ignoring methods with the right name but the wrong
signature, which is what would previously happen.

Also made the code more concise and "idiomatic".