Add a destructor method to objects.
Some objects may have "auxiliary storage", whose (de)allocation is not
handled by the runtime. This storage needs to be deallocated when the
object is freed. This cannot be done during the finaliser, as we want
the storage to be valid when other objects of the region get
finalised, hence this separate destructor method.

When a destructor runs, other objects in the region could have been
destructed and deallocated, making it unsafe to traverse any
references.

The VM will use this kind of “auxiliary storage” to store the array of
fields in objects, as well as string values.