Make the VM stricter about field writes
The STORE opcode is never valid on an IMM reference, so we can
disallow that. Additionally when writing a MUT value, the regions of
the two objects must match.

In practice the first check should never fail as the compiler will
disallow this statically. The second can happen as the compiler
doesn't enforce region equality yet, so it's nice to have a dynamic
check.