Modified finaliser to collect subregions
The finaliser must return the subregions that this object owns that will
require collection. This removes the need for the hacky
`trace_possibly_iso` that was used during deallocation.  This simplifies
the ordering of deallocation. The most general semantics the runtime
supports is
> self is mutable, but any mutable object it
> refers to is read-only

This implementation still allows the extract to function during
finalisation.  We may wish the language to be more restrictive, i.e.
everything read-only during finalisation.