Fixed test generation of CI
The test generation for CI is problematic.  When the CHUNK size was
made larger for the testing it meant the systematic test names no longer
match the regexp for making a smaller set of running tests in CI.

This test reduces the batch size in CI, so the original regexp still
works.