Simplify the CF stack and finalisers implementation.
- The current frame is stored at the top of the CF stack, rather than
  as a separate field.
- The current ip is stored in the top most frame, avoiding the need to
  save / restore it on call / return.
- Frames have a "halt on return" flag, used to make the return opcode
  in finalisers return control to the C++ code.
- Remove ident_. It can be inferred from the size of the cf stack.

This was prompted by a bug in the way finalisers used to be dispatched.
The receiver of the finaliser would be written to the last register of
the current frame. This could overwrite useful values, or underflow
and try to write to register 255, causing a crash. We were getting a
flaky CI because of it.