Improvements to systematic testing
Primary aim of this commit is to ensure the bug discovered while
reviewing 661ff54 does not reoccur. The code had originally not held the
epoch long enough. The fix to make ABA protection work correctly cross
platform also fixed the original bug on x86/x64 of not holding the epoch
before reading the tail of the SPMCQ. The bug did not occur in
systematic testing as the heuristics delayed "epoch ejection" until
enough work to make this required. The unit tests did not generate
sufficient work for this without long runs, and hence not easy to
explore in systematic testing.

This commit adds some extra interleavings into the systematic testing
and converts some on heuristics to coin flips during systematic testing.
The coin flips mean that we don't need very much work to exercise these
problematic paths.

It adds a test that exercises stealing more aggressively.