MLIR Prototype
This is the initial prototype for the MLIR lowering from Verona code.

It has the main features:
 * Creates an MLIR Generator, that transforms the AST into MLIR nodes.
   For now it can only lower one file, function.verona and does so in an
   opaque MLIR format (no dialect). It does not have a dialect yet.
 * A prototype of a dialect, empty, but with the minimal structure of an
   out-of-tree dialect to follow up when we start having proper Verona nodes.
 * A thin layer between AST and MLIR, to help convert the semantics between
   the two. It's far from complete, but it's an idea on how we should
   interface these two completely different representations.
 * Adds an LLVM build CI loop, where we cache the LLVM builds on Azure
   artefacts.
 * It uses the artefacts to test Verona without building LLVM from scratch.
   The artefact can also be used by developers on their machines via the
   same script the CI uses.

It has the main problems:

 * The Generator's public API is crude. It needs to account that we'll be
   using it to drive code through MLIR and LLVM, not always printing them
   to files.
 * Error handling is primitive. Uses the Error from ast for parsing the
   source files, but relies on throwing exceptions for the MLIR public
   API. This should be clearer once we have a better public API.
 * The Debug images for LLVM are too large for our Verona builder
   and we run out of disk. All other tests pass, including Windows Release,
   so I'm confident this is just a space issue. We need to decide how we're
   going to fix that.
 * The code is precariously generating opaque types and operations and not
   checking for anything. There are various asserts spread around to help
   convey meaning, but until we have a robust interface between AST and
   MLIR, and between the opaque MLIR and the Verona dialect, we won't know
   for sure what's the best way to lower to MLIR.
 * Lowering to LLVM only works when the MLIR is in standard dialect,
   which the opaque syntax is clearly not. We need to lower to Standard
   dialect first, then to LLVM, and that will take a while. I just left
   the code there as boilerplate.

Detailed comments from intermediary commits:

 - Initial Work

The main task here will be the AST->MLIR, a standalone tool, using the ast
and MLIR libraries to convert one to another. The actual
compiler, later on, will use all of the libraries in the same way, but
not call these tools directly. This way, we have individual tools for
testing as well as guaranteeing a more generic API for all libraries.

Interesting features:
 * Standard cmd-line options (`-o`, `--emit-llvm`, etc)
 * Auto-detection of file suffix (`.mlir`, `.ll`, etc)
 * Reads AST, MLIR, spits out MLIR

Adding IntegerType + storage

This commit adds a new base type, IntegerType, with parameters `width`
and `sign`, that can create different types. This is a proof of concept,
and an example for future types to come.

We don't need an opaque type, we just need to allow unknown types and
call them anything other than !verona<"...">, ex. !type.foo.

 - Parsing Verona function names and signature, body and return values

Finding function names from tokens, and argument and return types from
the signature, and lowering as an opaque "!type" type. We are being
very simplistic about types, especially the 'where' keyword that is just
treated as an alias, for now.

The current generator doesn't handle types at all, they're all opaque
and string based. There is a cast at the return to the function's return
type, as we're relying on the ast to have done the minimal checks, and
on the subsequent (to be implemented) type check to make sure it's
correct.

The operations are also opaque for now, function-like, returning single
values. This allows us to create unary and binary operators with varied
syntax ("verona.add", "verona.cast", etc). This flexibility helps us
generate code without caring much about the structure, which was checked
by the AST generator and will be validated by the lowering to Verona
dialect.

 - Reimplemented symbol table locally

The ScopedHashTable works well for symbols that cannot be updated, but
it doesn't allow us to redefine the value of symbols. This is bad
because we want to declare variables and later bind values to them.

 - LLVM submodule, build with artifact (Lin/Win/Mac)

Adding LLVM submodule on a specific commit and pulling the build
directory from DevOps so we don't have to build it on slow machines.
The scripts work on both Azure CI and desktop (Linux, Windows, Mac),
using Azure Blob Storage, which is trivial to get publi access with a
URL, so anyone (CI, developers) can use the same process.

Option to disable LLVM for Format build. May need something more robust
in the future.

 - Review comments and updates

A few updates after the pul request and review comments:
 * Documentation on the (temporary) ownership model of the arguments
 * Renames `NodeType` to `NodeKind` and 'Int' to 'NodeType'
 * Fixes typos on comments
 * Uses `uint8_t` for integer width
 * Fixes parser to understand `Snn` types as well (and add test)
 * Changed debug builds to RelWithDbgInfo to work around sizes
 * Adapted the tests to existing ctest structure