Initial commit for a cpp-peglib based parser and untyped AST gen (#116)
* Initial commit for a cpp-peglib based parser and untyped AST gen

This uses cpp-peglib to generate a packrat PEG for a grammar defined
in a text file, and then does some very basic AST manipulation. It
sets up scopes and symbol tables, and does a bit of fixity analysis.
The intention is to use this for a very slim parser front-end that
goes to MLIR before any serious heavy lifting is done. No tests are
included in the initial commit, as the grammar.peg is preliminary.Z

* use https submodule reference

* link threads

* add win32 path implementations

* extend path.h using get_builtin_library impl

* trying to pass ci

* windows and macos includes

* clangformat