DevOps LLVM Build (#147)
Builds LLVM, creates a tarball and uploads that as an artefact. This can
be used by the Verona build instead of rebuilding LLVM every time.

For now, the DevOps pipeline needs to set the $(GOOD_HASH) manually, but
the idea is to automate this later in the process for finding the new
LLVM commit that we want to rebase on top of.