Rewrite object hashmap
This PR makes several improvements over the initial implementation of
the hashmap. These improvements include documentation of member
functions and the internal algorithm used for the implementation. Tests
have also been added to compare the correctness of the map against the
standard `unordered_map` on randomized sequences of insert, update, and
delete operations.