Add LLVM's handling to building instructions
This is a temporary work around for trying to cache the LLVM build and is how it's working on the CI and some developers' machines already.

If we want more automation or seamless "cmake" magic, we may have to define what's the kind of caching we really want, long term, and work towards that, or we revert to always building LLVM on every Verona build and it will be much simpler (but a lot slower).

Also, on developer machines, building LLVM only happens once, but on the CI, because the machines are not on 24/7, we need to rebuild the whole LLVM, making a simple pre-commit test take hours.