Refactor MLIR generator to use ast_utils for all AST handling
This removes every dependency that the MLIR generator had on the
AST structure. All searches in the AST structure and accessors to names,
IDs, types are done via helper functions.

The only knowledge of the AST is via a NodeKind switch in parseNode,
which is a lot better than calling "isFoo" and "isBar" functons, but it
could be made so if people feel strongly inclined. Given that this is a
helper enum, not an AST enum, it should be ok.

The logic for accessing the AST structure is now 100% encoded in the
helper functions and comments were added where things were less than
clear. This makes it easier for us to exchange knowledge and refine the
API of both AST and MLIR to make things simpler.