Addind constant nodes
Constant nodes are, for now, special opaque verona operations with
opaque types given by their AST node tags. This is necessary as MLIR
initialises constants with the operation "constant" which need to be of
specific MLIR types, which are not the same as Verona types yet.

Once we have a complete type system and a map between Verona and MLIR
types defined, we can change this to emit attributes directly into
constant operations.