Adding support for function calls and return values
Changing how to detect basic-block termination to account for all known
terminators and empty basic blocks, so we don't need to terminate blocks
that have already been handled.

Implementing return values (of single expressions) as well as function
calls with the appropriate casts for the arguments, given we don't yet
know all types at this time.

If Verona accepts multiple return values, we'll need to change the
return value of all `parseX` functions to a vector of values. This
should be done later, as a refactor PR with examples. It will also fix
the `return mlir::Value()` problem, as we can return an empty list.