Use CMake ExternalProject for LLVM (#197)
Use CMake ExternalProject_Add to download or build LLVM.

By setting

-DVERONA_DOWNLOAD_LLVM=On
It will download the correct LLVM blob, and then build using that.

By setting

-DVERONA_DOWNLOAD_LLVM=Off
It will build the submodule as a subproject, and once that has completed build the Verona files.

The core complication is that Verona is using CMake files produced by the LLVM build, so we need to schedule the build
of Verona after LLVM has completed. This means we build Verona as a subproject, so that we can execute the cmake configure phase after the LLVM build has completed.