Add basic type checking of copy and return.
A new verifier is added to the copy op, requiring the input to be a
subtype of the output.

Similarly, a new `verona.return` op is added, allowing subsumption on
the returned value: it doesn't have to match the return type exactly,
but it should be a subtype of it.