Lowering loops as while op
Lowering while loops on the new `while` operation in the Verona dialect.
This is a prototype only, no variable tracking and accessing variables
exclusively via loads and stores (like before).

The next step is to change the dialect's `while` syntax to only use
arguments and return variables, to make sure we don't blindly capture
variables from outside the region, and thus can use the
`IsolatedFromAbove` trait that allows parallelism in the IR passes.

This syntax is mainly to make sure the control flow makes sense,
especially the loop_exit construct, receiving the condition of the loop
on the same basic block. We could have left it as the previous
conditional branch to a block with a `verona::BreakOp`, but that is
harder to infer semantics.

Also removing `currentFunc` as the region identifier and using insertion
block's parent as the region to create basic blocks. With nested
regions, that became necessary.