Adding loop_return operation
The need for a special loop_return is to implement function return inside
a region, as opposed to region return, which return values from the region
and need the region to have return values declared explicitly, and will
be used by value tracking later on.

For now, this is a loop-only return, to indicate bypass, but it could be
a general return (including function) if we need our returns to be
differentiated (ex. subtyping).

However, implementing an actual function return needs us to interface
with the ReturnLike trait and implement all its interfaces, and it's not
clear yet this is necessary (ie. subtyping can be done with casts).

Once the loop construct is lowered to CFG blocks, this operation will be
converted to a standard return.