Initial type checking of MLIR (#206)
A new typechecking pass is added, which the copy op hooks into, requiring the input to be a subtype of the output.

Previously the subtyping would stop as soon as pattern matching
succeeds. If a rule matched the right types but failed because of the
rule's specific logic, further rules would not be attempted.
However there wasn't one order of rules that would recognise all
patterns. It now tries all rules, until one of them works.

Additionally, this includes a bit of C++ magic to make the definition
of subtyping rules concise, by using the type of lambda parameters to
drive the pattern matching phase.

There are still some patterns, namely distributivity, that can't be
recognised yet. Supporting those needs a normalisation pass before the
syntax-directed rules are applied.

* Add a test for subsumption in copy and return.

* Prepare for CR.

* Add test output.

* Allow `verona.return` to take zero or more operands.

* Add missing copyrights.

Update the CI script to also include '.td' files in the check.

* Clang format.

* Remove the return operation.