Adding support for simple `for` loops
This has a separate implementation for for as for while, but the code is
really different. To unify we'd have to move on to a different strategy,
ie. build nodes from MLIR objects and have wrappers to get them from
AST.

This has uncovered a number of issues when calling functions, which will
be addressed in following patches.