Fix function return types
Stronger requirements on return values and return operations for
functions. The AST is still not rigid about the types so we need to emit
the error in the MLIR side.

Also implementing the TODO to move from 'type.none' to actual no return
type, getting rid of the 'none' cast and using simple 'return's.