[NFC] Add pragma once to ast-utils.h
The header has no guard and would break compilation if used twice.
However, we also need to move it to a static struct, as the functions
would be defined twice, and break linkage.

This wasn't a problem before, when the header only declared the
functions (and ast-utils.cc would define them), but since we moved
implementation to the header (due to the use of templates on array
filling functions), we now have all definitions in the header.

This is a non-functional-change, that just makes it safer to use the
ast-utils on other files, when needed.