[NFC] Refactoring function parsers/generators
Separate between parsing an AST function (decl/def) and generating MLIR
FuncOps, so that we can reuse the code on compiler-generated
functions.

This is the first of a series of changes that will split the generator
in two areas, and could potentially see the whole class split in two,
between AST parsing and MLIR generation.