WIP Refactor of C++ API for RT unit tests.
The C++ unit tests were using `new` to allocate Verona objects, but
were partially initialising the object using the Verona RT, and then
letting the C++ constructor fill in the rest.  On GCC, this was
triggering dead-store elimination for the Verona RT initialisation and
causing crashes.

This change makes a collection of static methods that build the Verona
object, with an embedded C++ object.

This commit only updates one of the unit tests as an experiment to see
how it looks with this alternative API.