Split backpressure bits into status and state
This change splits the `Backpressure` bits for each cown into two parts:
1. `BackpressureState` - may be modified from multiple threads, and therefore requires compare exchanges for modification
2. `Status` - storage bits for load heuristics,  token message tracking, and future I/O related things.

The large piece of documentation for backpressure has been removed, in favor of a complete rewrite. Some major changes are likely to come soon, and the documentation has diverged quite a bit from both the model and the implementation already. #226 has been updated to include new documentation.