Update backpressure progress
This brings the backpressure implementation up to date with the model,
where overloaded or unmutable message participants result in the other
participants becoming unmutable as well. This only applies when the fast
send is interrupted, since that is the only time a recipient may be muted
prior to running the message.