Alter CMake policy
`SNMALLOC_CI_BUILD` is an option to snmalloc, but CMake policy ignores
it being set on first pass, and then not
on second pass.  This change defaults to the new behaviour
where it is not ignored on any pass.
This should fix the double builds in CI of anything depending on the
runtime.