Refactor to store meta-data before Object*
GCC was applying dead-store-eliination to the region-meta data
as it was part of Object, and assigned before the C++ constructor.
This changes the representation of Object, to not contain the meta
data, but access the meta-data directly before the Object.

There are some defensive code paths to try to ensure that each Object*
actually has the meta-data before it.