Add flags to help with writing negative tests
Two flags are added to the verona-mlir binary to help with writing
negative tests, where we expect an error to be raised.

--verify-diagnostics allows error messages to be written inline within
MLIR files. The compiler succeeds only if all errors listed in the
file get emitted.

--split-input-file allows for a single input file to be split into
pieces, where each piece is treated independently. This works around
the fact that MLIR verifiers abort on the first error they find. By
splitting the input file we can include multiple related test cases
and expected errors in a single file.