Add support for class predeclaration
Allowing classes that haven't been declared yet to be predeclared and
then redefined later.

Refactoring `parseClassType` back into `parseClass` to avoid confusion
that we need two different type parsers. In a previous iteration I've
added that logic straight into `parseType` but there are too many
differences and the code was hard to follow. In theory, that should only
be done at the point if the class declaration, which is when
`parseClass` is called.