Refactoring how classes are lowered
Previously, we were treating the module as a special case of class and
lowering a (so far) redundant class representation. This was getting in
the way of simple lowering of methods and nested classes.

This refactory changes how this is done by treating the module as just
another class, and lowering sub-classes, methods and fields in a natural
way. There is no need for a class operation, as the class type is
lowered correctly and is associated with every function via a function
attribute.

Compiler-generated functions (ex. `for` loop's `next`/`has_value`) do
not have a class context. It is debatable if others will need it, and
the intention is to have them declared in Verona code soon enough, in
which case, they will have their classes as an attribute then.

Class types also have gained a "$parent" node, pointing to where they
were declared (either $module or some other class, when nested). This
may turn out to be irrelevant, but it's cheap to add and can help us
when trying to mangle the name of methods correctly.

Only the class test had to be changed, to remove the class operations
and add nested and static method tests. All other tests had their
function declarations changed to account for the new attribute (except
compiler generated functions).