Add new Verona UnknownType, remove opaque
Using opaque types won't validate with Verona operations, so we need an
unknown type that can validate. This is as much a place-holder as the
opaque "unk" but is part of the Verona type family.

We may want in the future to have two classes of types: with unknown and
without unknown, to help with validation of certain operations that
cannot take an unknown, but that's not for this patch.

This patch replaces all previously opaque types for Verona's unknown, to
make it easier for the type inference pass to only look at those and not
any existing opaque type.

It also completely removes the use of opaque types, introducing default
associations of int/float literal types from the AST. We may want to
de-sugar those types into actual Verona types in the AST, or after the
initial lowering (setting them to unk and letting type inference handle
literal type size adjustment).

For now, we need this to lower some Verona operations, so it will have
to do. Changes to the type lowering will become clearer once type
inference is defined.