Adding new operator for empty classes only
This adds support for the new operator, with tests for new and existing
regions, without initializers.

Done:
 * Empty class, new region
 * Empty class, existing region

The existing region points to the `alloca` of the new variable, which is
not great. The next change adds the class type to the alloca, so at
least it's clear to which alloca this is refering to.

We can only really fix this once we get rid of the spurious allocas, but
that can only trully happen when we also fix basic-block arguments for
conditions and loops, and in turn, that can only happen when we have
free variable analysis.

It's not yet clear if we want this to be done during lowering (and use
the AST information, like rengolin/zzz/mlir-freevar) or later, as a pass
to remove those, or as part of an existing pass.