Remove class operation from MLIR dialect
As discussed, this operation is redundant and serves no purpose in the
lowering of class operations. The type has all the information we need
to walk field names and types and can be used for field access,
allocation, etc.

This also simplifies a lot the implementation of `new` (in the following
patches), as there is no need to validate the types against both types
and field operations.