Change alloca type to arg/rhs type
Also, changes how to handle intrinsic functions, that now need a
compatible type for each application. This is not right yet, but we'll
probably have those being declared by the Verona builtin modules anyway,
so for now, they're a stop-gap.

Previously, all allocas were `!verona.alloca`, then `!verona.unk`, but
we can know what the type is by looking at the right-hand side (or the
function argument, for stack arg allocation) to infer what the type is.

This clarifies the IR, potentially simplifyinf the job of the type
inference pass, as well as allow classes allocated in others' regions to
check that the `alloca` is for the same variable.

We should get rid of so many allocas, but for now, the CFG on
conditionals and loops doesn't have free variable analysis, so we don't
know how to build the PHI / BB args lists. This is for another patch
series.