Adding field_read / field_write
Implementing field read/write for some simple cases. Not all options covered
yet. We have a few examples that show the mechanism working, but there
are a few cases that are not covered.

To make this work, the type checker had to be altered to accept unknown
types as "always pass". The rationale behind this is that we can't
really check unknown types before inference, and after inference, there
should be no unknowns.

Furthermore, we can make it a hard rule of any pass *after* type
inference that no unknown types can be added, even temporarily, to the
IR. This makes sure that we don't have to run additional clean up passes
where the pass' own cleanup didn't remove all instances.