Simplify Symbol Table Impl/API
Simplifying the implementation by only having one lookup and one
update/insert method and making "the other" a wrapper with a default
boolean flag.

Also simplifying the API by making all methods return the element if
inserted/updated/found and use that to avoid multiple redundant lookups
throughout the code.

The main difference is the use of `update` when declaring classes, so
that it is clear we can pre-declare them before they're fully defined
(for example, as a type for some earlier class' member).

Before that logic was hidden through a lookup + insert, and now it
simplifies as just an update.

In the end we have a lot less redundant map searches and a simpler code
to follow.