Refactor class/type lowering, add static calls
This is a massive change, but it's one that moves towards simpler code
and IR. The main changes are:

1. There are no special types any more

Integer, float, bool are all "classes" and they get casted to whatever
needed for arguments and return values. All numeric types will be coded
like classes with dynamic calls in the future, so that fits better with
the model.

Lowering of native arithmetic will be done when lowering to LLVM, with
special lowering for known patterns.

2. The DescriptorType was created

That's a handler to a class for static calls. Unlike dynamic calls, we
don't have an object to infer what the type is, so we create this type
that will hold a (global, unique) handler to lower a call in the same
way as dynamic calls, but with that handler instead of an object.

3. Two new operations were created: Call and Static

Call implements both dynamic and static calls, by taking the handlers
(descriptor or object) and call a method by name and arguments.

Static gets the descriptor of a type, to be able to call its static
methods.