Improve code quality around systematic logging
There are various log messages that are called that go to either the
flight recorder or the systematic logging. In Release builds, the
compiler should erase all the logging code, so it doesn't affect
performance.

This change makes that more true.  It removes various static
initialisers that were not optimised away by Clang.