Classes are modules
Using MLIR modules to represent class nesting and function ownership, to
be able to natively represent functions with the same name from
different classes without having to create a new operation and use the
ill-suited MLIR region for that.

Modules can be nested and merged, so this could make includes reasonably
trivial to make if we change the base module's name from "$module" to
the file name, or module name or some unique identifier.