Allow clearing many registers at once.
When variables go out of scope, the compiler emits Clear instructions
for each of them, which can become very verbose in run logs.

By having a separate opcode which accepts an arbitrary number of
registers we can reduce the noise (and shrink the bytecode size,
although that wasn't the main goal).