Switch the value stack from a vector to a deque.
When a register is cleared, this may cause a re-entrant execution on
the VM is order to run the finaliser. This could lead to stack
growing and being relocated in the middle of the Value::clear call,
giving very surprising results.

By making it a deque, we ensure growing the stack never causes Value
objects to move, keeping their address stable even in the event of a
re-entrant call.