Fix alloca values not being loaded
During the type refactoring, the comparison to load a value moved to
`unkTy` which was the wrong way of querying an alloca.

While the real fix is to not need those in the first place, they help
simplify condition/loop lowering. Once we simplify `for` loops, for
example, and we get to a single loop lowering, we should expand it to
carry free variables as arguments to basic blocks, and then we can get
rid of the alloca/store/load mode.

Most test changes are the introduction of loads before calls, but the
re-numbering of SSA values make it look like it's a big thing.