Reviving the freeVar analysis
Dynamic class, run on parseFunction, clear before the end.
simpler dumping, to stderr to not break tests.

Does not give any context on the CFG. May not be enough to get all
special cases like break and continue. But it could be a good way to
avoid having to go beyond the first predecessors and sucessors in the
CFG.