De-sugar for loops in AST + hygienic_id
Adding desugar ast pass with one transformation for replacing `for`
loops as `while` loops with the appropriate API calls. The types used in
the `for` loop are supposed to implement the iterator interfaces that
have methods: has_value, apply and next.

Some helper functions are added to sugar.cc, as it's not clear they will
be useful outside the de-sugaring context, but others were added to
ast.cc, which are available to all passes.

Those are:
 * insert_before(child, before): insert `child` into `before`'s parent,
   before that node.
 * move(ast, child): move `child` to its new parent, `ast`.
 * move_before(child, before): same as insert_before, but move.
 * move_children(ast, old): move all children from `old` into `ast` and
   clear `old` children.

Also, added a hygienic ID per class scope, auto incrementing a number
and allowing a prefix to help read the ast tree. The counter should be
module-wide, but given that the module is a class, and it's stil unclear
how we'll handle it, I'm just finding the nearest `classdef`. This could
be improved later.