Add debug flag to stop processing after passes
Adding a `-s|--stopat` flag to stop processing the ast after pass X. I
had to add a "gen" stage to return before any of the passes run, so that
we can see the raw ast, too. Not passing the flag or passing an empty
string will not stop at any pass.

Examples:
 $ verona-ast -a -s gen file.verona # prints the raw ast
 $ verona-ast -a -s ref file.verona # prints ast after the 'ref' pass
 $ verona-ast -a file.verona # prints the final ast tree, as seen by MLIR

With `-a` printing at the end, we can print the ast at any stage of the
parser, which and helps understanding the structural requirements of
each pass.

It's not possible to create tests yet due to the nature of our
(hard-coded) test runner (in CMake), but the output is just the tree
after a certain pass.