Replace backpressure state with priority (#310)
This is the first step towards implementing the full backpressure system, as described by the TLA+ spec. This is mostly cosmetic, except that prioritization via overloading is a state transition that may occur before a message is processed by a cown. Also, cowns are now lowered to a normal priority when their queue becomes empty and goes to sleep.