Implement `invoke` ast node in MLIR
Invoke is similar to `call`, where the first argument is a `member`
instead of a reference to an object. But the reference to the object is
in the member, together with the method's name.

Adapting the ast-utils methods to work around that made the MLIR
implementation identical and therefore unchanged. Added test for invoke
on existing file.