Use MLIR's ODS to generate Type definitions.
This cuts down on a lot of boilerplate around the definition of types,
accessors and storage classes.

Because of complications due to its recursive structure, we can't use
ODS with ClassType yet. We may be able to use ODS with a custom
storage class and extra class declarations, but it leads to a messy
Frankenstein definition, so it is easier to keep it all in C++.

Similarly, we can't use the parsing/printing code ODS generates,
because our implementations need extra context about recursive classes
(ie. the class_stack fields of TypePrinter and TypeParser).