Implement block arguments (PHI nodes)
This adds block arguments for conditionals and loops, and removes the
need for alloca/load/store special operations.

For conditionals, only exit blocks have PHI nodes, so it's easy to
use the free variable analysis from the ast to set those as arguments
and just use the symbol table at each point to know what values will be
operands.

For loops, we need to pass the arguments to all blocks (head, body, tail),
handle comtinue/break and update the symbol table on the lower scope just
like conditionals.

To simplify continue/break and arguments, I have created a new
structure, LoopFlowControl, which has head/tail blocks (for cont/brk)
and the list of arguments, by name, to search the symbol table when
needed (and update them to the block's arguments).

Those args are used by continue/break to pass the current (in scope)
values as operands to the branch instruction.